Add files via upload
White Background Handtap Theme originally made by Hende. I wanted a white background with a black hand, so I inverted the images in Photoshop. This inflated the size of the images, but it works with Substitute and Anemone 2.x, have not tested with Amemone 3. Does not work with Cydia Substrate and Snowboard.